Harden POL_Shortcut_InsertBeforeWine against special chars
Pass the command(s) to insert to the Perl script as a variable instead
of inserting them directly into the script; This prevents any
interpretation of special chars in the command by Perl.

Handle explicitly \n's

Make sure the commands are only inserted before the first occurence of
"POL_Wine".